Lack of additional info from pcre_study isn't fatal.
From pcre_study manpage:
If the function returns NULL, either it could not find any
additional information, or there was an error. You can tell the
difference by looking at the error value. It is NULL in first case.

Previously, Ag simply died if pcre_study returned NULL, without
checking to see which case had caused it.  This doesn't seem to
be a problem on OS X 10.8 with libpcre 8.31, but it causes
unnecessary failure on Debian 5 with libpcre 7.6.