search: Adjust log levels for max search depth
When a user has specified a specific max search depth (--depth), it is
reasonable to assume that they are aware of it. In this case, it should
suffice to output this information on the debug level.

Relatedly, it is not (in my opinion) an error to exceed the max search
depth, since the program is designed that way. I've lowered it to a
warning.

This increases, in my experience, the user friendliness of the --depth
option, since it outputs a lot less noise.