Allow stdin redirection
If `ag foo <file` is used, try to search the file like a file and not
like a stream. This has better performance and supports multiline
regexes. Fall back to stream behaviour if the system doesn't support
`/dev/stdin` or `/dev/fd/0` as this is still better than not supporting
it at all.