windows: allow non-win32-specific build and address warnings
on environments which can build ag for windows with the default Makefile, such
as MXE, print_w32.c was not built since it only appears at Makefile.w32, and so
the build failed.

include this file also at Makefile.am but put the whole of it in #idfef _WIN32.
also add correct includes, removed some unused variables, and initialize others