windows: make matches use color too
the code assumed that state doesn't carry between calls, and as such
treated the screen color at the time of call as the "reset" color, and
used this reset color when exiting the function.

it does seem like it at least tried to keep the state (static attr_old)
however it's never actually used as static and therefore overwritten on
each call. (same goes for static stdo which is overwritten every time,
but at least this one doesn't hurt us other than slighly slower code).

this broke the matches colors since the color is printed in one fprintf
call, and the text in another, and so the color was reset after the
first call and didn't carry into the second, so it ended up without color.

this patch initializes attr_old only once, removes the color reset at
the end of each fprintf, and also doesn't reset it before any color change.

also removes the static for stdo since it's never used as static.