windows: don't use bold color for everything
- plain 30-37 and 40-47 should not set bold on, but they did
- reset should (and now does) also reset all non color attributes
- blink now used for bg intensity (same as bash/less/ansicon for windows)
- uncommented and fixed fg/bg specific color resets
- replaced a magic number 6 (max sequence length) with something more readable
- removed 100 as reset (90-97 and 100-107 are aixterm codes for bright fg/bg)

it's still not as robust as other ansi color interpreters for windows,
especially when it comes to complex sequences, but now it does get the basics
mostly right and similar to others for simple sequences

also, fprintf_w32 tries to handle several other types of ansi sequences which
ag never uses (cursor position, etc) except for K which is typically no-op
anyway. I didn't look at those parts of the code, but they could probably be
removed safely.