Remove NOP option print-long-lines
The options --print-long-lines do not actually do anything.
The silver searcher's default behavior is to print long lines.
While it would be really nice if by default it did not and this option
was required to turn that on, that has not been implemented. Having
this option is misleading to folks that are looking for a way
to have ag remove the long lines from the output.