Add faster substring search
This commit implements the fast substring search algorithm described by
Leonid Volnitsky (see http://volnitsky.com/project/str_search/).

The idea is we insert each consecutive pair of characters in the needle
into a hash table. We then step through pairs of characters in the
haystack. If the pair is in the hash table (and also probably the needle)
we do a character-by-character comparison.

Performance is 30-40% faster than Boyer-Moore on 2GB of text:

> ./ag_hash --stats blahblahbla ~/Downloads/wiki/
0 matches
0 files contained matches
4 files searched
1956034523 bytes searched
0.800452 seconds

> ./ag_master --stats blahblahbla ~/Downloads/wiki/
0 matches
0 files contained matches
4 files searched
1956034523 bytes searched
1.310972 seconds