scandir.c: traverse in consistent order
`readdir()` returns directories in an implementation-dependent order.
In practice it's based on a hash of the filenames, aka random:

    $ ls
    01.txt  02.txt  03.txt

    $ ag -l hello
    02.txt
    03.txt
    01.txt

This commit calls `qsort()` on the output of `readdir()`.
Now, results are almost always sorted alphabetically:

    $ ag -l hello
    01.txt
    02.txt
    03.txt

This doesn't guarantee 100% sorted order - worker threads will still
race each other - but 99% of the time it will be sorted completely.

This means you can largely expect to get the same `ag` output on
different OSes or different checkouts. Plus, if you're running a slow
`ag` command and see the "foo" directory:

    $ ls
    ack  bar  foo  quux

    $ ag hello
    foo/hello.txt
    1:hello

    ...

you can be sure there are no matches in "ack" and "bar".

This commit partially fixes #107.