Add support for ~/.agrc
Using shell aliases for ag to change the defaults (such as --color-match)
is ugly and breaks trying to use ag nicely in cases like xargs.
As such, it's very helpful to have a default ".agrc" file which can be
read to apply custom options.

The .agrc file file should contain a list of command-line options, one
per line. Only lines which start with '-' are used, all others are
ignored.

The default agrc file is $HOME/.agrc, but can be overridden with
--agrc=<agrc-file>
--no-agrc or --noagrc are supported to disable reading from agrc.
There is no support yet for multiple agrc files.

Tests have also been updated to pass --noagrc to keep a user's defaults
from failing tests.