Fix incorrect path_start which can lead to heap buffer overflow
Fixes #1039
path_start is supposed to be the part of path which isn't in base_path,
but an incorrect assumption about a trailing '/' in base_path combined
with an off-by-one issue led to a buffer overflow of path.

This patch fixes the above problems by adding a trailing '/' to
base_path when necessary during parse_options() as well as strengthening
the logic for figuring out path_start.
Since path and base_path don't change in the middle of search_dir,
I further optimized by calculating path_start in search_dir once rather
than for every file in the directory in filename_filter.