Add --all-files option
This option causes ag to print headings for all files that it searches,
not just those that contain matches. This will be useful in the case
where ag is being called by another program that needs to show the user
a progress indicator of how many paths have been searched.

Signed-off-by: Max Brunsfeld <maxbrunsfeld@gmail.com>