win32: detect and use normal ansi on msys2/cygwin tty
Under msys2/cygwin tty, std* are pipes with specific name format. Detect
it as tty, and if win32 binary - use normal ansi colors on such case.

This allows a unified native win32 binary which can run either in
native windows console or inside msys2/cygwin terminal.

This commit is mostly no-op when building native msys2/cygwin binary
(where _WIN32 is not defined and ansi colors are by default).

But it does also improve native msys2/cygwin builds due to the new
isatty test when decising whether or not to search stdin - which is a
FIFO on msys2/cygwin env.