build: workaround when fopencookie exists but zlib or lzma do not
While zfile.c does have some #if HAVE_ZLIB_H and #if HAVE_LZMA_H - both
for the headers and in the code, if HAVE_FOPENCOOKIE==1 then zfile.c still
tries to unconditionally access zlib and lzma symbols, which causes the
build to fail (to test: './configure --disable-lzma --disable-zlib && make'
on a system which has fopencookie).

The correct fix is more involving and requires either refactoring zfile.c
or adding more than few #if tests for zlib and lzma throughout the code,
which would clutter it quite a bit. This can be improved later.

For now though, just don't use fopencookie if either zlib/lzma is missing.