refactor make_lang_regex to avoid potential memory overflow issues
As discussed in [1], the strncpy or memcpy in combine_file_extensions is
potentially risky because the extension is never actually checked
against SINGLE_EXT_LEN.  Furthermore, the split between
combine_file_extensions and make_lang_regex is unnecessary, as the
result of the former is only used to be passed to the latter.

Merge the logic of combine_file_extensions and make_lang_regex together
and skip the intermediate linear buffer step. Now make_lang_regex takes
the list of selected language indicies and produces the regex. This also
means that SINGLE_EXT_LEN can be removed, as well as the arbitrary limit
of 100 extensions to match against.

[1] https://github.com/ggreer/the_silver_searcher/issues/1378