Update RHEL/CentOS installation instructions
This updates the RHEL/CentOS installation instructions to resolve a number of
issues.  Most importantly, the command didn't work before unless you
already had epel-release installed.  It has to be installed before yum
can see the_silver_searcher, so it needs to be a separate command.
Another important detail is that epel-release is only available in the
default repositories of CentOS, not RHEL, so installing it by name
doesn't work out the box.  A better approach is to use the convenience
"latest" link provided by the EPEL project.  Finally, clarify that this
works on both RHEL and CentOS, and for version 6 and up.